+++ resources/lang/en/darkmode.php
@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 return [
     'light'        => 'Light',
     'dark'         => 'Dark',